requirements.txt
langchain
langchain-community
langchain-groq
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
langchain-huggingface

